Added VS project and solution. Changed LogentriesAppender to use the new unified ConfigurationManager for hosting-agnostic configuration.
