+++ src/LeLog4net.sln
@@ -1,0 +1,20 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 2012
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LeLog4net", "LeLog4net\LeLog4net.csproj", "{C20C9C59-A38E-4D01-B3DA-0E83013BD2EC}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{C20C9C59-A38E-4D01-B3DA-0E83013BD2EC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C20C9C59-A38E-4D01-B3DA-0E83013BD2EC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C20C9C59-A38E-4D01-B3DA-0E83013BD2EC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C20C9C59-A38E-4D01-B3DA-0E83013BD2EC}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
