+++ src/Packages/Microsoft.WindowsAzure.ConfigurationManager.2.0.1.0/Microsoft.WindowsAzure.ConfigurationManager.2.0.1.0.nuspec
@@ -1,0 +1,23 @@
+<?xml version="1.0"?>
+<package xmlns="http://schemas.microsoft.com/packaging/2011/08/nuspec.xsd">
+  <metadata>
+    <id>Microsoft.WindowsAzure.ConfigurationManager</id>
+    <version>2.0.1.0</version>
+    <title>Windows Azure Configuration Manager</title>
+    <authors>Microsoft</authors>
+    <owners>Microsoft</owners>
+    <licenseUrl>http://go.microsoft.com/fwlink/?LinkID=235167</licenseUrl>
+    <projectUrl>http://go.microsoft.com/fwlink/?linkid=252450</projectUrl>
+    <iconUrl>http://go.microsoft.com/fwlink/?LinkID=288890</iconUrl>
+    <requireLicenseAcceptance>true</requireLicenseAcceptance>
+    <description>Windows Azure Configuration Manager provides unified API to load the configuration settings regardless of where the application is hosted - whether on-premises or in a Cloud Service.</description>
+    <summary>Windows Azure Configuration Manager provides unified API to load the configuration settings regardless of where the application is hosted - whether on-premises or in a Cloud Service.</summary>
+    <releaseNotes />
+    <copyright />
+    <language />
+    <tags>Microsoft Azure Configuration Configuration Manager Cloud windowsazureofficial</tags>
+    <references>
+      <reference file="Microsoft.WindowsAzure.Configuration.dll" />
+    </references>
+  </metadata>
+</package>
