+++ src/Packages/log4net.2.0.0/log4net.2.0.0.nuspec
@@ -1,0 +1,23 @@
+<?xml version="1.0"?>
+<package xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
+  <metadata>
+    <id>log4net</id>
+    <version>2.0.0</version>
+    <title>log4net [1.2.11]</title>
+    <authors>Apache Software Foundation</authors>
+    <owners>Apache Software Foundation</owners>
+    <licenseUrl>http://logging.apache.org/log4net/license.html</licenseUrl>
+    <projectUrl>http://logging.apache.org/log4net/</projectUrl>
+    <requireLicenseAcceptance>false</requireLicenseAcceptance>
+    <description>NOTE: This package contains the Log4Net 1.2.11 assembly. log4net is a tool to help the programmer output log statements to a variety of output targets. In case of problems with an application, it is helpful to enable logging so that the problem can be located. With log4net it is possible to enable logging at runtime without modifying the application binary. The log4net package is designed so that log statements can remain in shipped code without incurring a high performance cost. It follows that the speed of logging (or rather not logging) is crucial. 
+
+At the same time, log output can be so voluminous that it quickly becomes overwhelming. One of the distinctive features of log4net is the notion of hierarchical loggers. Using these loggers it is possible to selectively control which log statements are output at arbitrary granularity. 
+
+log4net is designed with two distinct goals in mind: speed and flexibility</description>
+    <summary>The Apache log4net library is a tool to help the programmer output log statements to a variety of output targets. NOTE: This package contains the Log4Net 1.2.11 assembly.</summary>
+    <releaseNotes />
+    <copyright />
+    <language />
+    <tags>logging log tracing logfiles</tags>
+  </metadata>
+</package>
